+++ src/main/java/shagbot/guihelp/Launcher.java
@@ -1,4 +1,4 @@
-package shagbot;
+package shagbot.guihelp;
 
 import javafx.application.Application;
 
@@ -10,4 +10,5 @@ public class Launcher {
         Application.launch(Main.class, args);
     }
 }
+
